Moved Document to its namespace.
